+++ flutter_badge_manager/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.0.2
-- **ADDED**: Split to each supported platform
+- **ADDED**: Separation for each supported platform
 
 ## 0.0.1
 - **ADDED**: Initial Release